Move handlers and routes from app to controller
